fixet feed of learnrate with typecast to float32

+++ pretrainingYolo/pretraining.py
@@ -270,13 +270,13 @@ def main():
             #training:
             if i < 250000:
                 for j in range(nr_of_epochs_until_save_model):
-                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate)})
+                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)})
             elif i < 350000:
                 for j in range(nr_of_epochs_until_save_model):
-                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate/10)})
+                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)/10)})
             else:
                 for j in range(nr_of_epochs_until_save_model):
-                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate/100)})
+                    _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)/100)})
